OverlapReport moved to lib. possibly unneeded
